refactor: VO/Service/Validator class 분리
