+++ src/main/java/com/tech/baseball/BaseballGameView.java
@@ -1,8 +1,10 @@
+package com.tech.baseball;
+
 import java.util.Scanner;
 
 public class BaseballGameView {
 
-    private final Scanner scanner = new Scanner(System.in);
+    private Scanner scanner = new Scanner(System.in);
 
     //숫자 입력
     public String getUserNumber(){
